moving os upgrade state machine into package
