+++ machines/__init__.py
@@ -1,0 +1,1 @@
+from machines.os_upgrade_state_machine import OsUpgradeStateMachine
